requirements.txt
pandas
numpy
seaborn
pydentic
scikit-learn
mysql-connector
dotenv
-e .